requirements.txt
annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.1.1
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
distlib==0.3.7
fastapi==0.103.1
filelock==3.12.4
greenlet==2.0.2
h11==0.14.0
idna==3.4
nox==2023.4.22
packaging==23.1
platformdirs==3.10.0
psycopg2==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
virtualenv==20.24.5
